refactor: split GameAnimations into separate files under Animations directory
Each animation component (FightArrow, ReapAnimation) now lives in its own file,
with shared constants and resolvers extracted into dedicated modules.
