+++ client/Components/GameBoard/Animations/constants.js
@@ -1,0 +1,8 @@
+export const ANIMATION_DURATION = 0.8;
+
+export const FIGHT_DURATION = 0.7;
+
+export const AnimationType = Object.freeze({
+    Reap: 'reap',
+    Fight: 'fight'
+});
